pyproject.toml
[project]
name = "art-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "appdirs>=1.4.4",
    "appnope>=0.1.4",
    "beautifulsoup4>=4.14.3",
    "black>=26.3.1",
    "chump>=1.6.0",
    "cython>=3.2.4",
    "dill>=0.4.1",
    "einops>=0.8.2",
    "ffmpeg>=1.4",
    "fvcore>=0.1.5.post20221221",
    "huggingface-hub>=1.7.1",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "joblib>=1.5.3",
    "jsonlines>=4.0.0",
    "loguru>=0.7.3",
    "mmengine>=0.10.7",
    "networkx==3.2.1",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "optree>=0.19.0",
    "pandas>=3.0.1",
    "pycocotools>=2.0.11",
    "pyrender>=0.1.45",
    "pyrootutils>=1.0.4",
    "pytest>=9.0.2",
    "pytorch-lightning>=2.6.1",
    "rich>=14.3.3",
    "roma>=1.5.6",
    "scikit-image>=0.26.0",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "timm>=1.0.25",
    "wandb>=0.25.1",
    "webdataset>=1.0.2",
    "xtcocotools>=1.14.3",
    "yacs>=0.1.8",
    "power-spherical",
    "gdown>=5.2.1",
    "detectron2",
    "setuptools<70.0.0",
    "pyarrow>=23.0.1",
]

[tool.uv.sources]
power-spherical = { git = "https://github.com/nicola-decao/power_spherical.git" }
detectron2 = { git = "https://github.com/facebookresearch/detectron2.git", rev = "a1ce2f9" }

[tool.uv]
no-build-isolation-package = ["detectron2"]
